refactor(shim): PR-9.5 — 删 preprocess / presets_io / queue_io / task_snapshot 4 shim
batch 3 of 6 — top-level → services misc 4 个。

  删 studio/{preprocess,presets_io,queue_io,task_snapshot}.py（4 × 10 行）

17 caller 文件 Python 脚本批量更新（同 batch 2 模式）。

回归：1584 passed (2 pre-existing imagehash fail)。
