+++ studio/api/routers/projects/training.py
@@ -59,7 +59,8 @@ from ._shared import (
     _version_dir_or_404,
     _version_train_dir_or_404,
 )
-from .... import db, presets_io
+from .... import db
+from ....services.presets import io as presets_io
 from ....services.projects import jobs as project_jobs, projects, versions
 from ....services.dataset import scan as datasets
 from ....domain import RegAiConfig
